upload with bounding boxes
