+++ exts/edgeimpulse.dataingestion/data/icon.png
