+++ exts/edgeimpulse.dataingestion/data/preview.png
